SF #2794111: Fix integer overflow in unscaled mouse coordinates. Thanks, Martin Frost.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1891 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/include/mouse.h
@@ -101,8 +101,8 @@ struct mouse_struct {
 
   /* these are clipped to min and max x; they are *not* rounded. */
   short x, y;
-  /* unscaled ones, to not loose the precision */
-  short unsc_x, unsc_y;
+  /* unscaled ones, to not loose the precision - these need to be int to avoid overflowing 16 bits */
+  int unsc_x, unsc_y;
   /* coordinates at which the cursor was last drawn */
   short oldrx, oldry;
   /* these are the cursor extents; they are rounded off. */
